pyproject.toml
[project]
name = "foreign-whispers"
version = "0.1.0"
description = "Foreign Whispers: YouTube video dubbing pipeline (download, transcribe, translate, TTS, video stitching)"
license = { text = "LicenseRef-AGPL-3.0-only-with-Commons-Clause" }
requires-python = ">=3.11,<3.12"
dependencies = [
    "torch==2.1.0",
    "torchaudio==2.1.0",
    "TTS>=0.22.0",
    "yt-dlp",
    "youtube-transcript-api",
    "openai-whisper",
    "moviepy<2",
    "argostranslate",
    "streamlit",
    "fastapi>=0.115",
    "uvicorn[standard]",
    "python-multipart",
    "pydantic-settings",
    "pyrubberband",
    "librosa",
    "soundfile",
    "pydub",
    "requests",
    "setuptools<81",
    "pyyaml>=6.0.3",
    "logfire",
    "pyannote-audio",
]

[dependency-groups]
dev = [
    "pytest",
    "httpx",
    "nbconvert>=7.17.0",
]
alignment = [
    "logfire",
    "silero-vad",
    "pyannote.audio",
]

[tool.uv]
constraint-dependencies = [
    "onnxruntime<1.24",
    "numpy<2",
    "transformers<4.46",
]
find-links = [
    "https://download.pytorch.org/whl/cu121",
]

[[tool.uv.index]]
name = "pytorch-cu121"
url = "https://download.pytorch.org/whl/cu121"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu121" }
torchaudio = { index = "pytorch-cu121" }
torchvision = { index = "pytorch-cu121" }

[tool.pytest.ini_options]
pythonpath = ["."]
markers = [
    "requires_silero: needs silero-vad and torch installed",
    "requires_pyannote: needs pyannote.audio installed and FW_HF_TOKEN set",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["foreign_whispers"]